Combines:
- anit-lag: client-side prediction, extrapolation, quality/loot tier system,
  kill XP for ghost ships, NPC agent cleanup on sink, MAX_SIMPLE_SHIPS 200,
  2MB ships_json buffer, ghost AI player-proximity culling, new UI components
- opti: ghost spawn queue, fleet formation AI, world editor, admin ghost spawn
  API (load_ghost_spawns / ghost_spawns_to_json), ghost dormancy, wake radius

Conflict resolutions:
- ship_init.c: kept opti fleet/spawn-queue system; adopted anit-lag's
  player-based per-ship dormancy and sway constants (ghost_sway_phase)
- admin_api.c: kept opti's world-editor endpoints; added anit-lag's
  islands_reposition endpoint alongside; removed dead ghost_spawns_get/set
- admin_server.c: routes POST /api/ghost-spawns to opti's save handler;
  added POST /api/islands/reposition from anit-lag
- ship_init.h: merged function declarations from both branches
- server.c: changed ghost_spawns_load() call to load_ghost_spawns()

+++ server/src/net/network.c
@@ -121,7 +121,7 @@ int network_process_incoming(struct NetworkManager* net_mgr, struct Sim* sim) {
             // Unknown connection - might be handshake
             uint8_t packet_type = net_mgr->recv_buffer[0];
             
-            if (packet_type == PACKET_HANDSHAKE && received >= sizeof(struct HandshakePacket)) {
+            if (packet_type == PACKET_HANDSHAKE && (size_t)received >= sizeof(struct HandshakePacket)) {
                 const struct HandshakePacket* handshake = (const struct HandshakePacket*)net_mgr->recv_buffer;
                 network_handle_handshake(net_mgr, &from_addr, handshake, sim);
             } else {
@@ -153,7 +153,7 @@ int network_process_incoming(struct NetworkManager* net_mgr, struct Sim* sim) {
         
         switch (packet_type) {
             case PACKET_CLIENT_INPUT: {
-                if (received >= sizeof(struct CmdPacket)) {
+                if ((size_t)received >= sizeof(struct CmdPacket)) {
                     const struct CmdPacket* cmd = (const struct CmdPacket*)net_mgr->recv_buffer;
                     network_process_player_input(net_mgr, sim, conn->player_id, cmd);
                 }
